intégration nouvelle version composant PhfListe crée par Hervé
