--- Src/Liste/readme.md
@@ -1,3 +1,0 @@
-### PhfListe
-- [Wiki](https://github.com/PHF-Vetosoft/PHF_Manager/wiki/PhfListe)
-- [source](https://github.com/PHF-Vetosoft/PHF_Manager/tree/master/Src/Liste)
